bugfixing

+++ testing/gors.go
@@ -22,6 +22,7 @@ type gorServer struct {
 	phy    string
 	dbpath string
 	server *server.DBServer
+	closed bool
 }
 
 func (gs *gorServer) addr() string {
@@ -29,12 +30,13 @@ func (gs *gorServer) addr() string {
 }
 
 func (gs *gorServer) create(numChunks, redundancy int, verbose bool) string {
+	gs.closed = false
 	dbTestFolder := ""
 	if exists("/mnt/dbs/") {
 		dbTestFolder = "/mnt/dbs/"
 	}
 	gs.dbpath = dbTestFolder + "testDB" + fmt.Sprint(gorID)
-	gs.server = server.Start("", "127.0.0.1", 10000+gorID, numChunks, redundancy, gs.dbpath, 1024*1024*16)
+	gs.server = server.Start("", "127.0.0.1", 10000+gorID, numChunks, redundancy, gs.dbpath, 1024*1024*128)
 	gorID++
 	gs.phy = string("127.0.0.1" + ":" + fmt.Sprint(10000+gorID-1))
 	waitForServer(gs.phy)
@@ -42,12 +44,13 @@ func (gs *gorServer) create(numChunks, redundancy int, verbose bool) string {
 }
 
 func (gs *gorServer) assoc(addr string, verbose bool) string {
+	gs.closed = false
 	dbTestFolder := ""
 	if exists("/mnt/dbs/") {
 		dbTestFolder = "/mnt/dbs/"
 	}
 	gs.dbpath = dbTestFolder + "testDB" + fmt.Sprint(gorID)
-	gs.server = server.Start(addr, "127.0.0.1", 10000+gorID, -1, -1, gs.dbpath, 1024*1024*16)
+	gs.server = server.Start(addr, "127.0.0.1", 10000+gorID, -1, -1, gs.dbpath, 1024*1024*128)
 	gorID++
 	gs.phy = string("127.0.0.1" + ":" + fmt.Sprint(10000+gorID-1))
 	waitForServer(gs.phy)
@@ -55,9 +58,14 @@ func (gs *gorServer) assoc(addr string, verbose bool) string {
 }
 
 func (gs *gorServer) kill() {
-	gs.server.Stop()
-	time.Sleep(time.Millisecond * 10)
-	os.RemoveAll(gs.dbpath)
+	if !gs.closed {
+		gs.closed = true
+		if gs.server != nil {
+			gs.server.Stop()
+			time.Sleep(time.Millisecond * 50)
+			os.RemoveAll(gs.dbpath)
+		}
+	}
 }
 
 func (gs *gorServer) disconnect() {
